M20 phase 5: full kernel_main on x86_64 — shell prompt up (UP)
Wires the kernel core/drivers into the x86_64 build and reaches the
same shell prompt the i386 reference port shows.  This is the M20
DoD met: `make ARCH=x86_64` produces a bootable kernel that runs
through the full M1-M19+M18.5 boot sequence on a single CPU.

What this commit adds

Boot bridge:
- kernel/hal/x86_64/main_entry.c: x86_64_main_entry(uint32_t magic,
  uintptr_t info).  Called from boot.s long_mode_entry, validates the
  mb2 loader magic, translates the mb2 info struct into the mb1
  shape the rest of the kernel expects, then calls
  kernel_main(MULTIBOOT_BOOTLOADER_MAGIC, mb1_ptr).  Keeps kernel.c
  arch-independent.
- kernel/hal/x86_64/mb2.c: walks the mb2 tag stream, picks out tag
  types 4 (basic memory), 6 (memory map - repacked into mb1's
  size-prefixed entry format), and 8 (framebuffer), and stamps a
  static struct mboot_info in .bss with the equivalent mb1 flags.
  Unknown tags skipped silently.

Boot.s update:
- long_mode_entry now ends with `call x86_64_main_entry` (after the
  hello-world print kept as a sentinel for the serial log).  rdi =
  mb2 magic, rsi = mb2 info ptr per System V ABI.
- Multiboot2 header gains a type-5 framebuffer-request tag asking
  GRUB for 1024x768x32.  Without this, GRUB does not deliver a
  framebuffer tag and fb_terminal stays inert.

Stubs for deferred subsystems (UP, no APIC, no SYSCALL, no USB):
- kernel/hal/x86_64/m20_stubs.c: minimum-viable returns for
  lapic_*, ioapic_*, smp_boot_aps, smp_set_lapic_timer_count,
  syscall_dispatch, enter_user_mode_wrap, xhci_poll.  Lives in
  hal/x86_64/; shrinks as M20.5 (SMP+APIC), Phase 7 (SYSCALL/SYSRET),
  and a future USB port land real impls.

kernel.c gates:
- APIC bring-up + LAPIC-timer programming + smp_boot_aps blocks are
  now guarded by `#if defined(__i386__)`.  x86_64 stays on the 8259
  PIC for UP IRQ delivery (PIT IRQ0 fires fine via the legacy path).
  An `apic: x86_64 SMP/APIC bring-up deferred (M20.5)` line lands on
  the serial log when running on x86_64.
- kernel_main's mb_info arg widened from uint32_t to uintptr_t.
  Source-compatible on i386 (uintptr_t == uint32_t).  Matching
  widen in kernel/includes/multiboot.h + kernel/core/multiboot.c.

Makefile:
- x86_64 ARCH_C_SRCS gains mb2.c, main_entry.c, m20_stubs.c.
- x86_64 CORE_C_SRCS is now the full i386 list MINUS smp.c, lapic.c,
  ioapic.c, pci.c, syscall.c, virtio_blk.c, xhci.c, usb_hid.c,
  exfat.c — files that are either SMP/APIC-specific (M20.5),
  ring-3-specific (Phase 7), or rely on driver patterns that need
  64-bit revisits (block layer DMA, USB host).

Verified

i386 (`make`)
- Shell prompt up, all M19+M18.5 self-tests PASS (preempt,
  parallel on -smp 4, vmm, kmalloc, microbench).

x86_64 (`make ARCH=x86_64`)
- Hello-from-long-mode sentinel on serial.
- GDT (7 entries, 64-bit), IDT (256 entries), PIC remapped.
- Multiboot2 info translated, multiboot mb1 ptr handed to kernel.
- PMM: buddy ready, 254 MiB total free across DMA + NORMAL zones.
- VMM: 4-level paging active, 1 GiB identity via 2 MiB pages
  inherited from boot.s.
- Slab + kmalloc backend ready.
- VFS + ramfs + devfs + procfs + bcache all up.
- fb: 1024x768@32bpp mapped at 0xfd000000 (mb2 framebuffer tag
  delivered after the request-tag landed in the header).
- pit: 1000 Hz / 50-tick quantum running on 8259-delivered IRQ0.
- 6 modules registered + 1 driver, devfs populated, procfs populated.
- vmm self-test PASS, kmalloc self-test PASS, microbench 10 ms.
- preempt self-test PASS — kernel + hog interleaved across PIT
  ticks (hog ticks ~52M in 500 ms).
- vc: root VC 128x96 cells.
- shell `d-os>` prompt visible at log end.

M20-PROGRESS.md
- Phase 5 status flipped to shipped.  Phase 6 (PIT IRQ + scheduler
  + shell) folded into Phase 5 since it all came up together.
  Phase 7 (SYSCALL/SYSRET) explicitly deferred to M20.5.
- Lessons learned: mb2 framebuffer-request tag is mandatory for FB
  delivery, kprintf has no %l prefix (rolls forward as default and
  shifts va_arg consumption), mb2 walker uses 8-byte tag alignment,
  m20_stubs.c pattern for deferred SMP/USB symbols.
- Next concrete action updated to Phase 8 (docs + final commit).

+++ kernel/core/multiboot.c
@@ -40,10 +40,10 @@ static void print_u64_hex(uint64_t v) {
     else    kprintf("%x", lo);
 }
 
-int mboot_init(uint32_t magic, uint32_t info_ptr) {
+int mboot_init(uint32_t magic, uintptr_t info_ptr) {
     if (magic != MULTIBOOT_BOOTLOADER_MAGIC) return -1;
     if (info_ptr == 0) return -1;
-    g_mbi = (const struct mboot_info*)(uintptr_t)info_ptr;
+    g_mbi = (const struct mboot_info*)info_ptr;
     return 0;
 }
